requirements.txt
langchain
sentence-transformers
faiss-cpu
python-docx
pdfplumber
pandas
crewai
numpy
python-docx
python-pptx
langchain
langchain-community
openai
instructor
PyPDF2
python-docx
pandas
python_pptx
streamlit
